Fail gracefully for PHP versions older than 5.5
